+++ CHANGELOG.md
@@ -30,6 +30,7 @@
     - Elemental Resonance: Fervent Flames
     - Elemental Resonance: High Voltage
     - Elemental Resonance: Impetuous Winds
+    - Elemental Resonance: Shattering Ice
     - Wind and Freedom
 
 ## v0.3.dev0 (6 Aug 2023)